ABE53000/project/requirements.txt
picamera==1.13
flask==0.12.4
flask-restful==0.2.12
flask-cors==3.0.7
requests==2.18.4
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
imutils==0.5.1
redis==3.0.1
gunicorn==19.9.0